+++ perl/LinkedSpec/ActionIR/Contracts.pm
@@ -23,6 +23,7 @@ sub _require_lowering_deps {
   lower_return_general_statement => _require_dep($deps, 'lower_return_general_statement'),
   lower_return_imatch_statement  => _require_dep($deps, 'lower_return_imatch_statement'),
   lower_assign_method_statement  => _require_dep($deps, 'lower_assign_method_statement'),
+  lower_push_value_statement     => _require_dep($deps, 'lower_push_value_statement'),
   lower_regex_subst_statement    => _require_dep($deps, 'lower_regex_subst_statement'),
   lower_array_pipeline_expr      => _require_dep($deps, 'lower_array_pipeline_expr'),
   lower_if_flow_statement        => _require_dep($deps, 'lower_if_flow_statement'),
@@ -389,6 +390,18 @@ sub _build_passthrough_ir_contracts {
 sub _build_assignment_and_regex_contracts {
  my ($d) = @_;
  return [
+  {
+   id                 => 'push_value',
+   ir_node            => 'PUSH',
+   diag_name          => 'push_value',
+   unresolved_pattern => qr/\bpush_value\s*\(/o,
+   lower              => sub {
+    my ($code) = @_;
+    my $lower = $d->{lower_push_value_statement};
+    $code =~ s/\b(?<expr>push_value\s*(?<PAREN>\((?:[^\(\)\"\\']++|\"(?:\\.|[^\"])*\"|\'(?:\\.|[^'])*\'|(?&PAREN))*\)))/$lower->($+{expr}) || $&/ge;
+    return $code
+   },
+  },
   {
    id                 => 'assign_value',
    ir_node            => 'ASSIGN',